# bai-27-rest-api-1-2/api/docs/swagger.yaml
openapi: 3.0.0
info:
  title: Book Store API
  version: 1.0.0
  contact:
    email: nguyenphucnguyenvy@gmail.com
servers:
  - url: https://1x20l.sse.codesandbox.io/api/v1
    description: BookStore API for dev
...
paths:
...
  /users:
    get:
      security:
        - BearerAuth: []
      summary: get users 
      tags:
        - users
      parameters: 
        - in: query
          name: page
          schema: 
            type: integer
            default: 0
            example: 0
        - in: query
          name: limit
          schema:
            type: integer
            default: 3
            example: 2
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  users: 
                    type: array
                    items: 
                      $ref: '#components/schemas/User' 
                  pagination:
                    type: object
                    properties:
                      page:
                        type: integer
                      limit:
                        type: integer
                      length:
                        type: integer
        '404':
          description: not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          description: bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
    post:
      security:
        - BearerAuth: []     
      summary: add new user
      tags:
        - users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: 
                  type: string
                email: 
                  type: string
                  format: email
                imgUrl:
                  type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  newUser:
                    type: object
                    $ref: "#/components/schemas/User"
        '400':
          description: bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                example:
                  message: Invalid request
  /users/{id}:
    get:
      security:
        - BearerAuth: []
      tags:
        - users
      summary: get user by id
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: the user ID
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: 
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          description: bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
    patch:
      security:
        - BearerAuth: []
      tags:
        - users
      summary: update user
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: the book ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: 
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/Book'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          description: bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
    delete:
      security:
        - BearerAuth: []
      tags:
        - users
      summary: delete a user
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: the user ID
      responses:
        '200':
          description: OK
        '404':
          description: Not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          description: bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
...
components:
  schemas:
    Book:
      type: object
      properties:
        _id:
          type: string
        title:
          type: string
        description:
          type: string
        imgUrl:
          type: string
      example:
        _id: 5ea5700c9b99a306c9b0c268
        title: Places Search Box
        description: Fork from https://javascript.info/async-await
        imgUrl: http://res.cloudinary.com/learn-nguyenvy/image/upload/v1587900422/student/805afbd0e48202972948e82b0c2c20d0.jpg
        __v: 0
    User:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
        role:
          type: integer
          enum:
            - 0
            - 1
          default: 1
        isLogging:
          type: boolean
          default: false
        wrongLoginCount:
          type: integer
        avatar:
          type: string
      example:
        role: 0
        isLogging: true
        wrongLoginCount: 1
        _id: 5ea54ff9e3d8f9fe939a89b2
        name: nguyenvy
        password: $2b$10$X.wfHRypCk1Mb2si/jbEf.3/JipPtRH0utFr3YEqTRAtzcQ3ssqXq
        email: admin@gmail.com
        avatar: http://res.cloudinary.com/learn-nguyenvy/image/upload/v1587994173/student/1c287138f02a9fbe0abd01de94ab325b.png
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer